% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rollover_schedules.R
\name{create_rollover_schedules}
\alias{create_rollover_schedules}
\title{Create 'rollover' GPS fix schedules for Lotek PinPoint transmitters}
\usage{
create_rollover_schedules(
  tag_lifespan,
  time_zone,
  location_cycle,
  first_deployment_date,
  last_deployment_date,
  output_location
)
}
\arguments{
\item{tag_lifespan}{number of days tag is expected to last according to PinPoint Host software (this depends on fix schedule)}

\item{time_zone}{time zone where tags will be deployed}

\item{location_cycle}{time (in days) between GPS fixes}

\item{first_deployment_date}{first datetime that tags will be used}

\item{last_deployment_date}{last datetime that tags will be used}
}
\description{
Create a batch of 'rollover' GPS fix schedules that span many dates (say an entire field season).
}
\examples{
create_rollover_schedule(
tag_lifespan = 250,
time_zone = 'UTC',
location_cycle = 2,
first_deployment_date = '2019-01-20 18:00:00',
last_deployment_date = '2019-02-15 18:00:00',
output_location = '/Users/Jay/Desktop'
)
}
