% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_testing_schedules.R
\name{create_testing_schedules}
\alias{create_testing_schedules}
\title{Create GPS fix schedules to test Lotek PinPoint transmitters}
\usage{
create_testing_schedules(
  test_period = 12,
  time_zone,
  stagger_time = "5 min",
  fix_interval = "1 hour",
  start_datetime,
  transmitter_list,
  output_location
)
}
\arguments{
\item{test_period}{amount of time (hours) to test individual tags (12 is default as recommended by Lotek)}

\item{time_zone}{time zone where tags will be tested}

\item{stagger_time}{difference in time between tags for when they record fixes ('5 min' as recommended by Lotek)}

\item{fix_interval}{time between GPS fixes (default is '1 hour' as recommended by Lotek)}

\item{start_datetime}{datetime when tags will be first tested}

\item{transmitter_list}{list of transmitter IDs to be tested}

\item{output_location}{file path where GPS schedules will be saved}
}
\description{
Create a set of schedules to test Lotek PinPoint tags prior to deployment. By default, function
applies recommended fix schedules recommended by Lotek.
}
\examples{
create_testing_schedules(
test_period = 12,
time_zone = 'America/Boise',
stagger_time = '5 min',
fix_interval = '1 hour',
start_datetime = '2018-12-20 12:00:00',
transmitter_list = c('1745021', '1745026'),
output_location = '/Users/Jay/Desktop/'
)
}
