% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_asf_schedule.R
\name{parse_asf_schedule}
\alias{parse_asf_schedule}
\title{Read GPS schedule and parse ASF XML file structure}
\usage{
parse_asf_schedule(file)
}
\arguments{
\item{file}{path to ASF file generated using PinPoint Host software or this package}
}
\description{
Examine the structure of an ASF file used to program GPS fix schedules for Lotek PinPoint tags
}
\examples{
parse_as_schedule('.../2019-01-20_UTC_2-day_rollover.ASF')
}
