% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sched_gps_testing_fixes.R
\name{sched_gps_testing_fixes}
\alias{sched_gps_testing_fixes}
\title{Schedule GPS fixes for testing}
\usage{
sched_gps_testing_fixes(date_times, tz = "GMT", out_file = NULL)
}
\arguments{
\item{date_times}{either character or POSIXct vector of date times for which
to schedule an attempted GPS fix}

\item{tz}{character string indicating the \code{\link[base]{timezone}} of *input*
datetimes. These will be converted to GMT as required by the PinPoint tags.}

\item{out_file}{character path to output XML file containing the discrete rule
schedule. The .ASF extension need not be specified.}
}
\value{
writes an XML file (*.ASF extension) that can be loaded to a PinPoint tag
via the PinPoint Hose software from Lotek
}
\description{
This function generates a **testing** GPS fix schedule for Lotek PinPoint
tags in the appropriate .ASF XML format. The resulting file can then be loaded
to the tag via the PinPoint Host software (see that manual for details on
scheduling).
}
\examples{
Dependency for `create_testing_schedules()`
}
