% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_rollover_schedules.R
\name{view_rollover_schedules}
\alias{view_rollover_schedules}
\title{View rollover GPS schedules}
\usage{
view_rollover_schedules(
  tag_lifespan,
  time_zone,
  location_cycle,
  first_deployment_date,
  last_deployment_date,
  type
)
}
\arguments{
\item{tag_lifespan}{number of days tag is expected to last (provided by Lotek PinPoint Host software)}

\item{time_zone}{time zone where tags will be deployed}

\item{location_cycle}{time between fixes (in days)}

\item{first_deployment_date}{datetime when tags will first be deployed}

\item{last_deployment_date}{datetime last tags will be deployed}
}
\description{
Before creating schedules, ensure that Lotek PinPoint tags will record data at the proper times
by viewing fix schedule in a ggplot calendar or timeline
}
\examples{
view_rollover_schedules(
tag_lifespan = 250,
time_zone = 'UTC',
location_cycle = 2,
first_deployment_date = '2020-04-20 18:00:00',
last_deployment_date = '2020-05-15 18:00:00',
type = 'calendar'
)
}
