% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_time_zones.R
\name{view_time_zones}
\alias{view_time_zones}
\title{Look up time zones when creating GPS schedules}
\usage{
view_time_zones(region)
}
\arguments{
\item{region}{area where Lotek PinPoint tags will be deployed (e.g., country or UTC zone)}
}
\value{
list of time zones for a particular region
}
\description{
View a list of time zones that can be used with `lubridate` functionality
}
\examples{
view_time_zones(region = 'America')
}
